Move all functionality from lib/utils/erc20.ts and erc721.ts to other files
